Refactor `GlobalAllocator` to be safer `TrackingAllocator`
